requirements.txt
deepface == 0.0.79
Pillow == 10.1.0
numpy == 1.22.0
pyvirtualcam[numpy] == 0.9.1
sounddevice == 0.4.6
